+++ aws-distro-opentelemetry-node-autoinstrumentation/src/serviceevents/utils/index.ts
@@ -1,0 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { SEHHistogram, BUCKET_FACTOR, BUCKET_FOR_ZERO } from './seh-histogram';
+export { truncateString } from './data-sanitizer';
+export { getInstanceId, clearInstanceIdCache } from './instance-id';